+++ test/unit/fuzz-tools.test.js
@@ -107,6 +107,7 @@ import * as vascularV105 from '../../lib/vascular-v105.js';
 import * as vteV106 from '../../lib/vte-v106.js';
 import * as eddecisionV107 from '../../lib/eddecision-v107.js';
 import * as traumaV108 from '../../lib/trauma-v108.js';
+import * as traumaclassV109 from '../../lib/traumaclass-v109.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -158,6 +159,7 @@ const MODULES = {
   'vte-v106.js': vteV106,
   'eddecision-v107.js': eddecisionV107,
   'trauma-v108.js': traumaV108,
+  'traumaclass-v109.js': traumaclassV109,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];